+++ app/(drawer)/_layout.tsx
@@ -1,10 +1,11 @@
 import { Stack } from "expo-router";
+import { CustomHeader } from '@/components/CustomHeader';
 
 export default function DrawerLayout() {
   return (
     <Stack
       screenOptions={{
-        headerShown: false, 
+        header: () => <CustomHeader />
       }}
     />
   );
